🧹 Rename standard resolver to OS resolver. Reuse init functions in fs provider.
